+++ ETLBox/src/Definitions/Database/TableData.cs
@@ -1,7 +1,9 @@
-﻿using System;
+﻿using ALE.ETLBox.DataFlow;
+using System;
 using System.Collections.Generic;
 using System.Data;
 using System.Data.Common;
+
 
 namespace ALE.ETLBox {
     public class TableData : TableData<object> {
@@ -10,50 +12,66 @@ namespace ALE.ETLBox {
         public TableData(TableDefinition definition, int estimatedBatchSize) : base() { }
     }
 
-    public class TableData<T> : ITableData  {
+    public class TableData<T> : ITableData
+    {
         public int? EstimatedBatchSize { get; set; }
-        public IColumnMappingCollection ColumnMapping {
-            get {
-                if (HasDefinition) {
+        public IColumnMappingCollection ColumnMapping
+        {
+            get
+            {
+                if (HasDefinition)
                     return GetColumnMappingFromDefinition();
-                } else {
+                else
                     throw new ETLBoxException("No table definition found. For Bulk insert a TableDefinition is always needed.");
-                }
             }
         }
 
-        private IColumnMappingCollection GetColumnMappingFromDefinition() {
+        private IColumnMappingCollection GetColumnMappingFromDefinition()
+        {
             var mapping = new DataColumnMappingCollection();
             foreach (var col in Definition.Columns)
-                if (!col.IsIdentity)
-                    mapping.Add(new DataColumnMapping(col.SourceColumn, col.DataSetColumn));
+                if (!col.IsIdentity) {
+                    if (TypeInfo != null && !TypeInfo.IsArray)
+                    {
+                        if (TypeInfo.HasProperty(col.Name))
+                            mapping.Add(new DataColumnMapping(col.SourceColumn, col.DataSetColumn));
+                    }
+                    else
+                    {
+                        mapping.Add(new DataColumnMapping(col.SourceColumn, col.DataSetColumn));
+                    }
+                }
             return mapping;
         }
 
         public bool HasIdentityColumn => IDColumnIndex != null;
-        public List<T[]> Rows { get; set; }
+        public List<object[]> Rows { get; set; }
 
-        public T[] CurrentRow { get; set; }
+        public object[] CurrentRow { get; set; }
         int ReadIndex { get; set; }
         TableDefinition Definition { get; set; }
         public bool HasDefinition => Definition != null;
         int? IDColumnIndex { get; set; }
-
-        public TableData() {
-            Rows = new List<T[]>();
+        TypeInfo TypeInfo { get; set; }
+        public TableData()
+        {
+            Rows = new List<object[]>();
+            TypeInfo = new TypeInfo(typeof(T));
         }
-        public TableData(TableDefinition definition) : this() {
+        public TableData(TableDefinition definition) : this()
+        {
             Definition = definition;
         }
 
-        public TableData(TableDefinition definition, int estimatedBatchSize) {
+        public TableData(TableDefinition definition, int estimatedBatchSize)
+        {
             Definition = definition;
             EstimatedBatchSize = estimatedBatchSize;
-            Rows = new List<T[]>(estimatedBatchSize);
+            Rows = new List<object[]>(estimatedBatchSize);
+            TypeInfo = new TypeInfo(typeof(T));
         }
 
 
-        public string[] NewRow() => new string[Definition.Columns.Count];
         public object this[string name] => Rows[GetOrdinal(name)];
         public object this[int i] => Rows[i];
         public int Depth => 0;
@@ -64,7 +82,8 @@ namespace ALE.ETLBox {
         public byte GetByte(int i) => Convert.ToByte(CurrentRow[ShiftIndexAroundIDColumn(i)]);
         public long GetBytes(int i, long fieldOffset, byte[] buffer, int bufferoffset, int length) => 0;
         public char GetChar(int i) => Convert.ToChar(CurrentRow[ShiftIndexAroundIDColumn(i)]);
-        public long GetChars(int i, long fieldoffset, char[] buffer, int bufferoffset, int length) {
+        public long GetChars(int i, long fieldoffset, char[] buffer, int bufferoffset, int length)
+        {
             string value = Convert.ToString(CurrentRow[ShiftIndexAroundIDColumn(i)]);
             buffer = value.Substring(bufferoffset, length).ToCharArray();
             return buffer.Length;
@@ -82,53 +101,84 @@ namespace ALE.ETLBox {
         public int GetInt32(int i) => Convert.ToInt32(CurrentRow[ShiftIndexAroundIDColumn(i)]);
         public long GetInt64(int i) => Convert.ToInt64(CurrentRow[ShiftIndexAroundIDColumn(i)]);
         public string GetName(int i) => Definition.Columns[ShiftIndexAroundIDColumn(i)].Name;
-        public int GetOrdinal(string name) => Definition.Columns.FindIndex(col => col.Name == name);
-        public DataTable GetSchemaTable() {
+        public int GetOrdinal(string name) => FindOrdinalInObject(name);
+
+        private int FindOrdinalInObject(string name)
+        {
+            if (TypeInfo == null || TypeInfo.IsArray)
+            {
+                return Definition.Columns.FindIndex(col => col.Name == name);
+            }
+            else
+            {
+                int index = 0;
+                foreach (System.Reflection.PropertyInfo propInfo in TypeInfo.PropertyInfos)
+                {
+                    if (propInfo.Name == name) break;
+                    index++;
+                }
+                return index;
+            }
+        }
+
+        public DataTable GetSchemaTable()
+        {
             throw new NotImplementedException();
         }
         public string GetString(int i) => Convert.ToString(CurrentRow[ShiftIndexAroundIDColumn(i)]);
         public object GetValue(int i) => CurrentRow.Length > ShiftIndexAroundIDColumn(i) ? CurrentRow[ShiftIndexAroundIDColumn(i)] : (object)null;
 
-        int ShiftIndexAroundIDColumn(int i) {            
-            if (IDColumnIndex != null) {                
+        int ShiftIndexAroundIDColumn(int i)
+        {
+            if (IDColumnIndex != null)
+            {
                 if (i > IDColumnIndex) return i - 1;
                 else if (i <= IDColumnIndex) return i;
             }
             return i;
 
         }
-        public int GetValues(object[] values) {
+        public int GetValues(object[] values)
+        {
             values = CurrentRow as object[];
             return values.Length;
         }
 
-        public bool IsDBNull(int i) {
+        public bool IsDBNull(int i)
+        {
             if (Definition.Columns[ShiftIndexAroundIDColumn(i)].AllowNulls)
                 return CurrentRow[ShiftIndexAroundIDColumn(i)] == null;
             else
                 return false;
         }
 
-        public bool NextResult() {
+        public bool NextResult()
+        {
             return (ReadIndex + 1) <= Rows?.Count;
         }
 
-        public bool Read() {
+        public bool Read()
+        {
             IDColumnIndex = Definition.IDColumnIndex;
-            if (Rows?.Count > ReadIndex) {
+            if (Rows?.Count > ReadIndex)
+            {
                 CurrentRow = Rows[ReadIndex];
                 ReadIndex++;
                 return true;
-            } else
+            }
+            else
                 return false;
         }
 
         #region IDisposable Support
         private bool disposedValue = false;
 
-        protected virtual void Dispose(bool disposing) {
-            if (!disposedValue) {
-                if (disposing) {
+        protected virtual void Dispose(bool disposing)
+        {
+            if (!disposedValue)
+            {
+                if (disposing)
+                {
                     Rows.Clear();
                     Rows = null;
                 }
@@ -137,11 +187,13 @@ namespace ALE.ETLBox {
             }
         }
 
-        public void Dispose() {
+        public void Dispose()
+        {
             Dispose(true);
         }
 
-        public void Close() {
+        public void Close()
+        {
             Dispose();
         }
         #endregion